requirements.txt
notion-client
streamlit
pandas
deepl
python-dotenv